Move FileErrors and FileInfo
