+++ src/renderer/context/ProjectFilesContext.tsx/ProjectFilesContext.ts
@@ -2,7 +2,7 @@ import { createContext } from "react";
 import { TrackItemId } from "../../../common/Flavors";
 import { Take } from "../../../common/project/Take";
 import { TrackItem } from "../../../common/project/TrackItem";
-import { FileInfo } from "../FileManagerContext/FileInfo";
+import { FileInfo } from "../../services/fileManager/FileInfo";
 
 interface ProjectFilesContextProps {
   saveTrackItemToDisk: (take: Take, trackItem: TrackItem, blob: Blob) => Promise<void>;
