+++ src/renderer/services/fileManager/FileInfo.ts
